--- UnitTestGradeManager/UnitTestGrade.swift
@@ -1,12 +1,0 @@
-//
-//  UnitTestGrade.swift
-//  UnitTestGradeManager
-//
-//  Created by kakao on 2022/12/29.
-//
-
-import XCTest
-
-final class UnitTestGrade: XCTestCase {
-
-}
